requirements.txt
django
djangorestframework
psycopg2
django-storages
boto3
